requirements.txt
python-dotenv>=1.0.0
requests>=2.31.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
google-genai>=1.0.0
pydantic>=2.0.0
pytest>=8.0.0
httpx>=0.27.0
